+++ test/crossfile/hover.lua
@@ -746,3 +746,18 @@ hover = {
     name  = 'food',
     description = "I'm a multiline comment\n"
 }}
+
+TEST {{ path = 'a.lua', content = '', }, {
+    path = 'b.lua',
+    content = [[
+---@return string # 'this is a tab `\t`'
+local function <?f?>() end
+]]
+},
+hover = {
+    label = [[
+function f()
+  -> string]],
+    name  = 'food',
+    description = "@*return* — this is a tab `\t`"
+}}