+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## 1.21.0
 * `NEW` setting: `completion.showParams`
+* `NEW` `LuaDoc`: supports multiline comments
 
 ## 1.20.5
 `2021-4-30`