+++ changelog.md
@@ -3,6 +3,7 @@
 ## 1.21.0
 * `NEW` setting: `completion.showParams`
 * `NEW` `LuaDoc`: supports multiline comments
+* `NEW` `LuaDoc`: tail comments support lua string
 
 ## 1.20.5
 `2021-4-30`