+++ changelog.md
@@ -1,5 +1,10 @@
 # changelog
 
+## 1.21.0
+* `NEW` setting: `completion.showParams`
+* `NEW` `LuaDoc`: supports multiline comments
+* `NEW` `LuaDoc`: tail comments support lua string
+
 ## 1.20.5
 `2021-4-30`
 * `NEW` setting: `completion.autoRequire`